backend/requirements.txt
# Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Security
werkzeug==3.0.1
bcrypt==4.1.1

# Payment
stripe==7.8.0

# Server
gunicorn==21.2.0
gevent==23.9.1

# Development
black==23.12.0
flake8==6.1.0
